requirements.txt
Flask==3.0.0
flask-limiter==3.5.0
pymongo==4.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
Authlib==1.3.0
openai==1.12.0
requests==2.31.0
google-generativeai==0.3.2
groq==0.4.2
